shift flag functions into the flag text package
